Change ClosingBlock to ExpireHeight
